cleaned up main by moving controller-specific functions to their respective files
